backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
websockets==13.1
apscheduler==3.10.4
pytz
